update resmax, incepmax, in separate files, also update training script for different conv types of inception
