+++ timm/models/__init__.py
@@ -80,6 +80,8 @@ from .xcit import *
 from .HMAX import *
 from .alexnet import *
 from .ALEXMAX import * 
+from .RESMAX import *
+from .RESMAX_Inception import *
 
 from ._builder import build_model_with_cfg, load_pretrained, load_custom_pretrained, resolve_pretrained_cfg, \
     set_pretrained_download_progress, set_pretrained_check_hash
